feat(events) : put event item function into a new class for best structure of code
